// src/types/permission.ts
export enum Permissions {
  VIEW_DASHBOARD = 'viewDashboard',
  // USER MANAGEMENT
  MANAGE_USERS = 'manageUsers',
  MANAGE_ROLES = 'manageRoles',
  MANAGE_ADMINS = 'manageAdmins',
  // PLATFORM SETTINGS
  MANAGE_CRYPTO = 'manageCrypto',
  MANAGE_CASINO = 'manageCasino',
  MANAGE_SETTINGS = 'manageSettings',
  // BONUSES & REWARDS
  MANAGE_BONUSES = 'manageBonuses',
  MANAGE_ROYALTY_TIERS = 'manageRoyaltyTiers',
  MANAGE_CASHBACKS = 'manageCashbacks',
  MANAGE_REFERRAL_REWARDS = 'manageReferralRewards',
  MANAGE_TIER_AFFILIATES = 'manageTierAffiliates',
  MANAGE_WAGER_RACES = 'manageWagerRaces',
  // GAME MANAGEMENT
  MANAGE_GAMES = 'manageGames',
  // TRIVIA MANAGEMENT
  MANAGE_TRIVIA = 'manageTrivia',
  // PROVIDERS MANAGEMENT
  MANAGE_OPERATING_PROVIDERS = 'manageOperatingProviders',
  // PROMOTIONS MANAGEMENT
  MANAGE_PROMOTIONS = 'managePromotions',
  // PAYMENTS & TREASURY
  MANAGE_PAYMENTS = 'managePayments',
}